Added restart command
